[breaking-change] Factor out "display_7seg" resource.
With the expectation that "display_9seg", "display_14seg" and
"display_16seg" will be forthcoming.

There are no provisions in this resource itself for multiplexing
the display. It is expected that on boards with multiplexed displays,
an additional resource "display_7seg_ctrl" will be provided, and it
would have either an n-bit signal "sel" (binary encoded) or an n-bit
signal "en" (one-hot encoded). (This may be revisited in the future.)

Fixes #17.